pyproject.toml
[project]
name = "fastai-forum-mcp"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "requests",
    "rank-bm25",
    "mcp[cli]",
    "sentence-transformers",
    "faiss-cpu",
    "huggingface_hub",
    "anthropic",
    "scipy",
    "einops>=0.8.2",
]
